+++ odoo/local-src/cosanum_contact/models/__init__.py
@@ -1,1 +1,3 @@
 from . import res_partner
+from . import customer_value
+from . import marketing_type