+++ odoo/local-src/cosanum_crm/__manifest__.py
@@ -1,21 +1,19 @@
 # Copyright 2019 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 {
-    'name': 'Cosanum Contact',
-    'summary': 'Specific Contact features for Cosanum',
+    'name': 'Cosanum CRM',
+    'summary': 'Specific Sale features for Cosanum',
     'version': '12.0.1.0.0',
-    'category': 'Sales',
+    'category': 'CRM',
     'author': 'Camptocamp',
     'license': 'AGPL-3',
     'depends': [
-        'base',
+        'cosanum_contact',
+        'crm',
     ],
     'website': 'http://www.camptocamp.com',
     'data': [
-        'views/contact_view.xml',
-        'views/customer_value_view.xml',
-        'views/marketing_type_view.xml',
-        'security/ir.model.access.csv',
+        'views/crm_lead_view.xml',
     ],
     'installable': True,
 }