+++ HISTORY.rst
@@ -26,6 +26,11 @@ latest (unreleased)
 * BSCOS-52: Add Company Group
 * BSCOS-89: Add vendor_id field on Helpdesk
 * BSCOS-76: The description field on Helpdesk should be Html
+* BSCOS-60: Install partner-conact
+* BSCOS-60: Install partner-contact
+* BSCOS-62: Add new fields and models to res.partner
+* BSCOS-90: Display new suffix fields in opportunity
+* BSCOS-61: Override name_get for res.partner
 
 **Bugfixes**